[ADD] product_uom :module added
In this task , create new tab(inventory) in contact module.
where we can add one or more products and uom for customers. When product is selected, uom should be filtered to show only other uom possible in the uom category.
When the customer is selected, and lines are added the product should display the uom from customer